# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API quản trị hệ thống Weaviate Vector Database.",
        "title": "Weaviate Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/weaviate-admin",
    "paths": {
        "/backups": {
            "post": {
                "description": "Yêu cầu Weaviate đóng gói snapshot toàn bộ dữ liệu ra backend cấu hình sẵn",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Tạo bản Backup",
                "parameters": [
                    {
                        "description": "Thông tin Backup",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/backups/{id}/restore": {
            "post": {
                "description": "Yêu cầu Weaviate phục hồi dữ liệu từ một bản backup (snapshot) đã tồn tại",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Phục hồi bản Backup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID của bản Backup cần phục hồi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin Backend lưu trữ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.RestoreBackupRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cluster/nodes": {
            "get": {
                "description": "Trả về trạng thái của các node trong Weaviate cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Lấy danh sách Nodes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collections": {
            "post": {
                "description": "Khởi tạo một Collection mới",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Tạo Collection",
                "parameters": [
                    {
                        "description": "Tên Collection",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collections/{name}": {
            "get": {
                "description": "Xem chi tiết schema (vectorizer, properties) của một Collection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Lấy thông tin Collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tên Collection",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collections/{name}/objects": {
            "get": {
                "description": "Trình duyệt các bản ghi có trong Collection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "summary": "Xem dữ liệu Objects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tên Collection",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Giới hạn số lượng kết quả (Mặc định: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/collections/{name}/tenants": {
            "get": {
                "description": "Lấy danh sách toàn bộ tenants thuộc về một Collection",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Xem danh sách Tenants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tên Collection",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Đăng ký một tenant mới vào collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Thêm Tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tên Collection",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Thông tin Tenant",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.AddTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Gọi ping và liveness/readiness probe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Kiểm tra trạng thái Weaviate",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PingResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/main.PingResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AddTenantRequest": {
            "type": "object",
            "properties": {
                "tenant_name": {
                    "type": "string",
                    "example": "tenant_01"
                }
            }
        },
        "main.CreateBackupRequest": {
            "type": "object",
            "properties": {
                "backend": {
                    "type": "string",
                    "example": "filesystem"
                },
                "backup_id": {
                    "type": "string",
                    "example": "backup-2026-04-17"
                }
            }
        },
        "main.CreateCollectionRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Article"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "main.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "main.PingResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Weaviate đang hoạt động khỏe mạnh"
                },
                "status": {
                    "type": "string",
                    "example": "PONG"
                }
            }
        },
        "main.RestoreBackupRequest": {
            "type": "object",
            "properties": {
                "backend": {
                    "type": "string",
                    "example": "filesystem"
                }
            }
        }
    }
}